+++ src/test/java/org/owasp/dsomm/metricCA/analyzer/deserialization/YamlReaderTest.java
@@ -1,4 +1,4 @@
-package org.owasp.dsomm.metricca.analyzer.yamlDeserialization;
+package org.owasp.dsomm.metricca.analyzer.deserialization;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
