--- src/test/java/org/owasp/dsomm/metricCA/analyzer/yamlDeserialization/ActivityTest.java
@@ -1,14 +1,0 @@
-package org.owasp.dsomm.metricca.analyzer.yamlDeserialization;
-
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.owasp.dsomm.metricca.analyzer.deserialization.activity.Activity;
-
-@ExtendWith(MockitoExtension.class)
-public class ActivityTest {
-
-  @InjectMocks
-  private Activity activity;
-
-}
